requirements.txt
fastapi
uvicorn[standard]
joblib
scikit-learn
pandas
xgboost
PyJWT
passlib[bcrypt]
python-multipart
python-dotenv
motor
skforecast
tensorflow
numpy
pickle-mixin
aiofiles
category_encoders
dill
skforecast
xgboost
nltk
google-generativeai
openpyxl
python-multipart